refactor: 컴포넌트 분리에 따른 스타일 정리
